updated save method in Fog::Rackspace::BlockStorage::Volume and Fog::Rackspace::BlockStorage::Snapshot to skip creating cloud reources if identity was already set Fixes #1402

+++ tests/rackspace/models/block_storage/snapshot_tests.rb
@@ -3,18 +3,25 @@ Shindo.tests('Fog::Rackspace::BlockStorage | snapshot', ['rackspace']) do
   pending if Fog.mocking?
 
   service = Fog::Rackspace::BlockStorage.new
-  volume = service.volumes.create({
-    :display_name => "fog_#{Time.now.to_i.to_s}",
-    :size => 100
-  })
+  begin
+    volume = service.volumes.create({
+      :display_name => "fog_#{Time.now.to_i.to_s}",
+      :size => 100
+    })
 
-  volume.wait_for { ready? }
+    volume.wait_for { ready? }
 
-  options = { :display_name => "fog_#{Time.now.to_i.to_s}", :volume_id => volume.id }
-  model_tests(service.snapshots, options, false) do
-    @instance.wait_for { ready? }
+    options = { :display_name => "fog_#{Time.now.to_i.to_s}", :volume_id => volume.id }
+    model_tests(service.snapshots, options, false) do
+      @instance.wait_for { ready? }
+    
+      tests('double save').returns(true) do
+        @instance.save
+      end    
+    end
+
+    volume.wait_for { snapshots.empty? }
+  ensure
+    volume.destroy if volume
   end
-
-  volume.wait_for { snapshots.empty? }
-  volume.destroy
 end
